stdint.h instead of cstdint for older g++ compatibility. Define for collatz solver internal integer, so int64_t can be used to avoid int overflows. int64_t not needed for contest and int is faster (as contest states no soln's will overflow).

+++ FastCollatz.h
@@ -11,12 +11,10 @@
 // includes
 // --------
 
-#include <cstdint>  // int64_t
-
 /**
  * @param p the problem integer
  * @return the cycle length of p
  */
-int collatz_plen (int64_t p);
+int collatz_plen (int p);
 
 #endif // FastCollatz_h
